Fixes Bug#696375: x-image-meta-size is not optional, contrary to documentation.
The image's size is set to zero now during reservation of the image ID if the
image's size is not passed in with headers. In addition,
glance.store.Backend.add() now returns a tuple of (location, size) and
the image's size attribute in the registry is updated to this value if 
previously set to zero.

Adds a new test case that ensures the size attribute is set properly
when not included in the image meta headers.

Adds documentation for the new _reserve(), _upload(), and _activate()
methods in glance.server.Controller.