Fixes LP Bug #700162: Images greater than 2GB cannot be uploaded using glance.client.Client.
This patch introduces chunked-transfer encoding to the client
classes. When the glance.client.BaseClient.do_request() method
is called and the method is either POST or PUT and the body
argument is a file-like object, then the do_request() method
builds a chunked-transfer encoded request and send()s chunks of
the body file-like object over the httplib connection.

Not sure how we fully test this, since we'd have to pack up
a very large file for use in testing, but I'm open to suggestions :)