This adds a test case for LP Bug 704854 -- Exception raised by Registry server gets eaten by API server.
The test involves spinning up an API server and a registry
server, firing cURL and glance-upload against the API
server with invalid requests, and verifying that appropriate
exception messages are contained in the response.

I cannot reproduce 704854 now. Not sure what changed in
the code that actually fixed this, but the test case verifies
that the behaviour described in the bug does not occur any more.