unit/test_config.py: make it independent on sys.argv
The first test in test_config.TestConfig.test_parse_options tests
no options are specified. But it checks sys.argv when default parameter
is used. So it fails when some parameter is passed to run_test.py.
So pass empty list to the parser as argument explicitly in order to
make it pass independent sys.argv.