Fixes Bug #851216
Fixes a recurring issue with tests that were
decorated with the @runs_sql decorator where the
file-based SQLite registry database used in those
tests was not being properly cleaned up, which was causing
some random-looking failures when python setup.py test was
run for the DEB builders on Jenkins.

According to Monty, files needed in testing do indeed need
to be in MANIFEST.in, so I added the test_migrations.conf
file to that. Sorry, BrianW!

Change-Id: I6535b999921d159de5d957f6b70df3c13310e449