Fixes LP Bug #833285
Large objects, uploaded to Swift using the large object
manifest method, are uploaded in chunks/segments in the
Swift store driver.

Unfortunately, doing a DELETE on the object only deletes
the manifest, leaving the chunks/segments orphaned. This
patch makes the Swift store driver check for the existence
of an object manifest, and if found, deletes all intermediate
chunks, then deletes the object manifest itself.

Adds a functional test to check the segments are actually
deleted.

Change-Id: Ib5787581a57e419393faca198d867db757fc03b3