Fixes LP Bug#885341 - Test failure in TestImageCacheManageXattr
This bug was due to insufficient checking that not only was
python-xattr installed, but also the device that the image
cache dir lives in is actually able to have xattrs written
to files (user_xattr in fstab). The fix adds a number of checks
for xattr support on the device, both in the image cache configure()
methods and in the test cases themselves (using @skip_if_disabled)

Adds a functional test case for queue/prefetch of images into
the image cache.

Change-Id: I09b3d008fc00564031b0534155430ed0856e9e9c