Fixes LP Bug#837817 - bin/glance cache disabled
This patch removes the cache management commands
from the main bin/glance client and puts them in
a bin/glance-cache-manage program. It also adds
a number of cache management calls to manage the
cache queue via the HTTP API.

The reason I moved the cache management commands
into a separate utility is that I'm thinking towards
the future, where we have a separate admin API, and
having a separate management utility seemed like the
best way to go.

Change-Id: Ic916a72f39516a06be27ea6c8ef69eb9e1a70936