Bug#894027: use correct module when building docs
Currently the wrong directory is searched for the
glance module, requiring glance to be installed
in the default system path, for docs to be buildable.
Also the path was appended, so the default system version
which may be completely different is referenced.
This ensures that only local glance files are referenced.

Change-Id: Ia226c90bbd24d7804401a7c9e444387c78a0f542