Add new cfg module
As discussed on the mailing list and described here:

  http://wiki.openstack.org/CommonConfigModule

The module implements an API for defining configuration options and
reading values for those options that a user may have set in a config
file or on the command line.

The module will be part of openstack-common and glance will consume
it from there once openstack-common makes a release with an API
compatibility commitment.

Change-Id: Ideedf77e5e16fa4e6ca742cfb49f5539c5a4a746