Backport tox support for CI.
Add tox.ini file.

The tox.ini drives the current test suite in multiple virtual envs, so this
is a first step in trying to get multi-version testing.

Prep tox config for jenkins builds.

Add a generic tox build environment.

Allow for the running of arbitrary commands inside of a venv
so that we don't have to make tox envs for every blessed thing we want to
do.

Align to jenkins tox patterns.

Split requires in to pip-requires and test-requires.
Updated install_venv to know about test-requires.

Update tox.ini for jenkins.

The tox documentation is wrong, it looks for [tox:jenkins] instead
of [tox:hudson].  Also, run "pip freeze" before running tests from
within the virtualenv so we have a record of all versions used
in each test run in Jenkins.

Fix typo in tox.ini.

My last commit message was right, the code was wrong.  It's tox:jenkins.
Commands are invalid in tox: sections, so we'll get pip freeze run
another way.

Change-Id: I9c28ee11aa7bbe1b221a72e0ee8cc51d788f73c9