Fix help output for inverse of BoolOpt.
When a BoolOpt is used, it adds both the option and its inverse.  For
example, there is "--verbose", and "--noverbose".  Previously, the same
help text was used for both which didn't make any sense.

This patch modifies the help text used for the inverse option.  It now
simply says something like "The inverse of --<option>".  For example,
the help text for --noverbose says "The inverse of --verbose".

Change-Id: I3966fe9ec5ed6282d8a2902f06fc1487532064e0