Set size metadata correctly for remote images.
Previously the size metadata was always zero'd for remote images,
which was misleading and led to issues like:

  Bug 900959

We now query the remote store (HTTP, S3 or Swift) with the
equivalent of a HTTP HEAD in order to determine the correct
image size without downloading.

Where the size is available, the metadata is set appropriately.
Otherwise it falls back to zero as before.

Change-Id: I3093eba1b2fa023348558c45febee39e68e1a08f