Change xattr usage to be more broadly compatible.
This patch updates the xattr image cache driver to use methods that are
more broadly compatible.  The issue here is that there is more than one
xattr Python module out there.  Specifically, there are:

    https://github.com/xattr/xattr
    http://git.k1024.org/pyxattr.git/

The previous glance code expected an interface available in the first
version.  This patch changes the code to use an interface that is common
to both of these.

This makes all of the image cache tests working against the current pyxattr
package in Fedora.

Change-Id: I1fca66cfe914364c921408b676a06387d1de5c2a