blueprint progressbar-upload-image
The goal of this blueprint is to make cli more informative.
Adds optional animated progressbar upload to the Glance CLI tool.
This animation is spawned by a separated thread and the following information will be shown:

1. running progressbar
2. percentage of data has been transferred
3. bandwidth
4. ETA for transfer completion
5. status such as uploading or stalled

The animation can be disabled by adding the flag --silent-upload to the command "glance add"

Change-Id: If954268c6d7b52c975f6c358134be4f587541aa8