disk/container_format required on image activate
Fixes lp 937216

The initial change[1] to require the container and disk formats be
explicitly set was enforced on the initial image POST, regardless
of whether the image content accompanied that request (either in
as the entity-body, or externally specified via x-image-meta-location).

This turned out to be overly restrictive, as these formats may not
be known to the caller until the image content is supplied via a
subsequent PUT on the new image.

Now we only enforce the strict requirement for these formats to be
non-empty when the image is activated (as opposed to the initial
enqueue).

[1] https://github.com/openstack/glance/commit/62c913c3

Change-Id: I89e068f35fd7da427b547b18cdea4ae84ab3ec87