Glance skip prompting if stdin isn't a tty
Don't have glance issue prompts if stdin isn't a tty.  Fixes bug 884116.

In glance's user_confirm routine, check if stdin is a tty and if
not, just return the default value. If it is a tty, then continue
with the raw_input call to issue the prompt and get the response.
Latest patch includes recommended change to check for isatty attr before
using it.

Added myself to Authors.

Change-Id: I8134668b8382b44618c86972190f3d7d98f55e52