Support owner paramater to glance add
Fixes bug 962998

Allow the owner to specified as a first class image attribute
when creating a new image via the CLI, as it is currently for
image update.

Also added a simple fakeauth pipeline to allow the user, tenant
and admin status of API requests to be easily controlled by
functional tests.

Change-Id: I0f83d55d4a96ad3632fb238ad1758ec3f00ed3fd