Add MySQLPingListener() back
The use of wrap_db() to catch low-level problems in the connection
pool appears to be incompatible with new versions of sqlalchemy
(0.7.4/precise).  This re-instates the use of a event listener to
catch the error of 'mysql has gone away'.  Keeps the wrapping
logic in place for use before the connection pool has been constructed,
specifically to allow glance to retry its initial database connection on
service startup.

Fixes bug 967887

Change-Id: Ia732377d3404104b8acf7fe4ada164ec7ecf705f