Support for directory source of config files
Allow multiple config files to be pulled in from a config directory,
as opposed to individual config files being explicitly enumerated.

This logic is enabled using the --config-dir=/path/to/config CLI option,
causing config to be retrived from all matching /path/to/config/*.conf
files.

Sections may be re-opened across config files, and all config items
must reside in an explicitly specified section (i.e. it does not default
to [DEFAULT]). This behavior is unchanged.

The common elements of this patch (i.e. not specific to the
glance-specific location of paste.ini files) have landed in
to openstack-common as:

  https://github.com/openstack/openstack-common/commit/fd8c2e7c

Change-Id: I3b659f1837c082bd1a78fe860a468f9d68f0f132