Fix content type for qpid notifier.
Fix bug 980872.

This patch fixes a regression I introduced in
2d36facf14f4eb2742ba46274e04a73b5231aece.  In that patch, I adjusted the
content_type for messages sent with the qpid notifier to be
'application/json' to match a change that went into the kombu notifier.
Unfortunately, it's wrong.

I assumed based on the kombu change that notifications were being json
encoded before being passed into the notification driver.  That's not
the case.  The message is a dict.  So, just revert the change to set the
content_type and let Qpid encode the notification as 'amqp/map'.

Change-Id: Iea027409f6200109b97fed93073cff84bc915536