Fix glance-api process leak in respawn storm test.
Fixes a glance-api process leak that occurs when running
TestMultiprocessing.test_interrupt_avoids_respawn_storm.

The test code creates an array of pids sorted numerically sorted in
reverse (largest to smallest) meaning the child pids should be at the
top of the list. This patch simply grabs the child pids from the top of
the list instead of relying on the length of the list.

Fixes LP Bug #1000931.

Change-Id: I91d538e686a770778d9855a1bd5a1767cc83842e