Move RequestContext.is_image_* methods to db layer
These methods need direct access to the db and therefore require
the request context module to maintain a connection to the database.
Moving them to the db layer allows us to further refactor the
db layer.

This does duplicate the methods in each of the db drivers, so a
future patch is needed to simplify that.

Related to bp refactor-db-layer

Change-Id: Ia7e0dbb9f2162fb69837699b970f8cab3efcd545