Avoid thread creation prior to service launch.
It appears that the thread serving out the fake HTTP endpoint used
in the copy-from tests is causing intermittent hangs in the launched
glance service processes. These hangs manifest as calls to futex(2)
that never return, and occur almost daily in Smokestack.

The background to the thread/fork mismatch is explained in this
blog post:

 http://rachelbythebay.com/w/2011/06/07/forked

We avoid the issue by delaying the thread creation until after the
service processes have been launched by the testcase.

Change-Id: I68c6f1735e5001641994ed4f84ad879397ba8713