Combine same-time tests with main db test case
Rather than re-run all of the base test methods to test with
image fixtures with identical created_at timestamps, we should
just run methods that matter. This patch removes the subclass
that sets identical timestamps in favor of adding a third fixture
to the main test case.

Related to bp refactor-db-layer

Change-Id: I69d0b21d1ef825333881e021dfa9610f1bbad334