Store swift images in separate containers.
Updates the swift storage backend so that when using multi tenant
each images gets stored in a separate container.

As part of this change the unit tests were updated to generate unique
swift container names for each test. This acts as a container prefix
in multi-tenant mode so the tests were also updated to cleanup all
swift containers matching the container name prefix as well.

The motivation for this change is so that we can set swift read/write
permissions on each individual image.

Partially implements blueprint: swift-tenant-specific-storage.

Change-Id: I43f33594fb40dee33ef52e8ae873722e20b09a10