Match path_info in image cache middleware
The url matching code in the image cache middleware should only look at
the unconsumed path, not the full path. Webob denotes this by using
'path_info' instead of 'path.'

This also adds some tests for the version negotiation middleware to ensure
it is returning a sane path for the image cache middleware to match.

Fixes bug 1026222

Change-Id: I522e2210e2b7272f9bb47602510a2910fa193415