Do not cache images that fail checksum verfication
On an image GET, recalculate the image checksum as the image
data is streamed to the client.  Verify that the checksum matches
the original checksum calculated when the image was added to Glance.
If checksum validation fails, purge the image from the cache.

This type of situation could occur if the backend image store
is malfunctioning.

bug 1028496

Change-Id: I9f38bac8360016bb12b5edaad87c50939a538cc0