Handle images which exist but can't be seen.
If the master and the slave glance have different permissions, you
can end up with a 409 exception at upload. For example:

 - the image is visible to the replication user on the master
 - the image is present on the slave, but not visible to the
   replication user

The replicator checks for the existance of the image by doing a HEAD
on http://server:port/v1/images/uuid and gets a 404 back. It therefore
decides the image doesn't exist and that it should be replicated. It
does a POST for that URL, and gets a 409 back saying that UUID is
already in use.

Resolves bug 1030396.

Change-Id: Id5fdb7977692593fd6c2e203f0440d4aa5339d92