Add swift_store_admin_tenants option.
Add a swift storage option to configure a default list of admin
tenant IDs which will be granted write access on any swift
images.

This setting might be useful when a user choose not to use a global
swift admin role when managing glance images. This might also
be useful when implementing delayed delete for glance swift
multi-tenant.

Partially implements blueprint: swift-tenant-specific-storage.

Change-Id: Id65e1c8d114c9551b99f6a18c527ba26aefd341f