Omit deleted properties on image-list by property
If a user updates an image property with purge-props, it is marked
as deleted. However, the query to filter by custom property ignores
this, and returns match even for records with deleted set. This patch
introduces a small change by inserting deleted=False in the query to
ignore those deleted properties.

Also, some changes in simpledb is introduced to handle user defined
properties in the correct format.

- Inserts deleted=False on properties filter query. Fixes bug 1025697.
- Updates simpledb to handle user defined property in correct format.

Change-Id: I435751ef6fc7ba0537e573fac54c34b81098271b