Allow admins to share images regardless of owner.
Updates the Glance DB's is_image_sharable function so that
an admin can share an image regardless of the context ownership.

This fixes issues when trying to write Glance functional tests
which use the noauth middleware (UnauthenticatedContextMiddleware).

The motivation behind this change is I'd like to write some
functional tests for image-members that work with noauth.

Fix bug 1021054

Change-Id: I6d939876a4c4bf5f191463f5ca434fc524a6ac59