Update rbd store to allow copy-on-write clones
This is in preparation for cinder blueprint
efficient-volumes-from-images. If the same rbd cluster is used by
glance and cinder, a new volume can be created as a copy-on-write
clone by the ceph cluster rather than downloading all the data from
glance.

Since bp api-v2-store-access is implemented, make the location for the
rbd store more specific, so cinder can check that it's accessing the
same backend as glance. fsid is a unique identifier for a Ceph cluster.

Within a cluster, the full specification of an rbd image includes
pool, image name, and optionally snapshot. Creating a copy-on-write
clone is only possible in rbd on a protected snapshot, so after
uploading an image to rbd, create a snapshot and protect
it. Protection means it can't be deleted until it is unprotected, so
you can't accidentally render any clones unusable by removing their
parent. Add a new exception, InUseByStore, to handle a snapshot or
image that's still protected or otherwise in use.

More details can be found at
http://ceph.com/docs/master/dev/rbd-layering/.

For backwards compatibility, only use these new features if they are
supported by the installed librados and librbd python bindings. Otherwise
fallback to the previous location specifier which only includes image
id, since cloning won't work anyway.

Since the location parsing is now a bit more complicated, add unit
tests for it.

Change-Id: I9711702276d966f88284fae44f9531b992cdd06f