Update restrictions on allowed v2 image properties
The list of allowed and readonly image attributes has gotten out of
date, so this change updates those lists. This change also adds a
similar check to ensure certain special properties names are reserved as
they would conflict with existing database properties that are either
not presently exposed or exposed under a different name.

Change-Id: I56d6c938dd6a5353e15b3e5ef913fae5b4629d52