Add policy enforcment for v2 api.
This patch adds policy checks in the images controller for all actions that need
to be enforced. It also fixed an issue with unit tests where the v1 unitests
were testing less than they should be. This was due to the default policy being
overridden in the test and causing a forbidden to be raised whenever policys
were changed, regardless of the policy rule added.

Fixes bug 1036846

Change-Id: Ib351d8c1e13164d02b02685c808b30045f7850ea