multi_tenant: Fix 'context' is not defined error
Corrects the variable name 'context' to be 'self.context' in the
configure_add method of the swift store. We changed/moved this code
late during the review process of the multi_tenant swift feature
and obviously didn't add a functional/unit test on the new location
which now causes this exception to get thrown:

    NameError: global name 'context' is not defined

This patch also includes a new glance function test to startup
the Glance API in multi-tenant mode. The test_multi_tenant_mode
test will fail with the previous swift store implementation.

Fixes LP Bug #1037745

Change-Id: I7681ad704b169c5fe815adf2fe704e5d18ea8e83