Raise Forbidden exception in image_get
The simple db driver now raises 'exception.Forbidden' when a caller
tries to access an image that is not owned by the provided context.
The list of images returned by image_get_all will also be filtered
to match what the sqlalchemy driver is expected to return.

Related to bp glance-simple-db-parity.

Change-Id: Idb2b85252d5ad33194d3bd8b5d8a9bc368968bd0