Add missing columns to migration 14
This adds created_at, updated_at, and deleted_at to database
migration 14.

Background Info!

If a new database is created at version 14, it doesn't run any
migrations scripts. It lets sqlalchemy create the database from a
set of sqlalchemy models.

When an existing database is upgraded to version 14, it will use
the migration script. What's important to note is that this is
currently impossible due to bug 1042925.

While it is typically bad practice to alter a migration script, this
is a special case as nobody could have used it. What this patch does
is complete the migration script in preparation for a future patch
down the road to fix the bug referenced above.

Fixes bug 1043041

Change-Id: I6dc2afa735fb2f82df71d58b10c1e1530fce2f89