Add kernel/ramdisk_id, instance_uuid to v2 schema
Nova depends on a few properties to be able to interact with
images in Glance effectively. The three added in this patch
are kernel_id, ramdisk_id and instance_uuid. These properties
are just being added to the default v2 schema, so they can
be overwritten if a consumer decides they need to.

Fixes bug 1039818

Change-Id: I5cfd5f0b5c8d1d526d56d99de738f9271aa641c1