Rewrite Swift store functional tests
The existing functional tests for the Swift store depend on
spinning up a full system. We really just want to test that
our store interface is good, so this patch adds tests for
just that.

Prevent swift store from prepending 'http[s]://' to auth url
if it is already there. This should be broken into a different patch.

Running the tests works the same way, just define a local config
with the necessary swift_* options. Set the GLANCE_TEST_SWIFT_CONF
environment variable to the location of that config.

The old full-stack tests have been removed.

Change-Id: I506993ce44354950534443c2101f8417a2a918e0