Fallback to a temp pid file in glance-control
Fixes bug 1046593.

Previously if the pid file was not writeable by the current user
then glance-control bombs out, *after* spawning the required service
but *without* caching the process ID anywhere.

Now we fall back to a temporary file, so that at least user can supply
that file to the corresponding glance-contol stop command.

Change-Id: I6c1daf36221f731fdb5a8bb59d3f91a595337af8